PBAP: Apply listStartOffset in the final list.
This change will apply the offset value in the searched
list of vcard values prepared after comparing with the
search values instead of applying the ListStartOffset on the
original list of all phone-book enteries. Without this patch
offset was being applied on wrong list, resulting in sending
the unwanted vcardlisting to Client which was violating the
PBAP specification.

This also consists of change which will allow the client to give
a substring instead of the whole contact name while doing search
by name and in return server will return all the contacts with
names starting with that substring.

Remove BODY header from PSE's response to Phonebook size request when
max list count is set as 0 from PBAP Client. Without this there
is PBAP Spec violation where body header should not be sent when
max list count is set as 0 while requesting for PB download.

Change-Id: I36d757a12d0cc94628d4f4ad91c9370397035c4e
CRs-Fixed: 725842, 785824, 554222